fix Warning

+++ 05-SceneManager/Camera.cpp
@@ -7,8 +7,8 @@ CGame* game = CGame::GetInstance();
 Camera::Camera(int mapW, int mapH)
 {
 	Cam_x = Cam_y = 0;
-	Width = game->GetBackBufferWidth();
-	Height = game->GetBackBufferHeight();
+	Width = (float)game->GetBackBufferWidth();
+	Height = (float)game->GetBackBufferHeight();
 	mapwidth = mapW;
 	mapheight = mapH;
 }
@@ -23,14 +23,14 @@ void Camera::SetPosition(float& cam_x, float& cam_y)
 {
 	Cam_x = cam_x;
 	Cam_y = cam_y;
-	Cam_y -= game->GetBackBufferHeight() / 2;
+	Cam_y -= (float)game->GetBackBufferHeight() / 2;
 	float ScreenY = mapheight - Height + ScreenH;
 
 	if (cam_y > ScreenY) {
 		Cam_y = cam_y;
 	}
 
-	Cam_x -= game->GetBackBufferWidth() / 2;
+	Cam_x -= (float)game->GetBackBufferWidth() / 2;
 	if (Cam_x < 0)
 	{
 		Cam_x = 0;
@@ -40,7 +40,7 @@ void Camera::SetPosition(float& cam_x, float& cam_y)
 		Cam_y = 0;
 
 	if (Cam_x >= mapwidth - Width - ScreenH)
-		Cam_x = mapwidth - Width - ScreenH;
+		Cam_x = float(mapwidth - Width - ScreenH);
 	if (Cam_y > ScreenY)
 	{
 		Cam_y = float(mapheight - Height + ScreenH);
